api/requirements.txt
python-dateutil==2.8.2
requests==2.31.0
python-dotenv==1.0.0
